Fix manpages and wifi bug
- Manpages can now be opened without a problem. Previously, they could
not be opened since the section was left in the name and the "man"
command could not understand something like "fork (2)"

- Fix a bug with wifi selection. nmcli uses a * to designate the network
to which we are currently connected to. If that network is selected then
awk would read that as the first column. On the other hand, if another
network was selected, awk would read in its name as the first column(as
it should do!).